+++ CHANGELOG.md
@@ -4,5 +4,7 @@
 
 - Moved +mip and mip.m in source tree to matlab/+mip and matlab/mip.m
 - In import.m, determine location of packages directory relative to location of current script file
+- Allow installing multiple packages at once via `mip install package1 package2 ...`
+- Allow uninstalling multiple packages at once
 
 ## [0.1.2] - 2025-11-22